// com.mobi.web/src/main/resources/public/app/workflows/models/workflow-execution-activity.interface.ts
/*-
 * #%L
 * com.mobi.web
 * $Id:$
 * $HeadURL:$
 * %%
 * Copyright (C) 2016 - 2025 iNovex Information Systems, Inc.
 * %%
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 * 
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 * 
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 * #L%
 */

import { WorkflowStatus } from './workflow-status.type';

/**
 * Interface representing a workflow execution activity object.
 */
export interface WorkflowExecutionActivityI {
    executionId: string;
    executorIri: string;
    executorUsername: string;
    executorDisplayName: string;
    startTime: Date;
    endTime: Date;
    succeeded: boolean;
    status: WorkflowStatus;
    isLatestActivity: boolean;
}

/**
 * Interface representing a workflow execution activity object used for table
 */
export interface WorkflowExecutionActivityDisplayI extends WorkflowExecutionActivityI {
  executionIdLabel: string;
  startTimeLabel: string;
  runningTimeLabel: string;
  details?: string;
  logs?: string;
}